# docs/openapi/hermes-v1.yaml
openapi: 3.1.0
info:
  title: Hermes Remote Admin API
  version: 1.0.0
  description: |
    Legitimate remote device administration — MVP.
    Admin routes require JWT (2FA verified). Device routes use long-lived device tokens.
servers:
  - url: https://api.example.com/api/v1
    description: Production (replace with your VPS hostname)
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Admin login (step 1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: Partial JWT — must complete 2FA
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginResponse'
        '401': { description: Invalid credentials }
  /auth/2fa/verify:
    post:
      tags: [auth]
      summary: Verify TOTP (step 2)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorVerifyRequest'
      responses:
        '200':
          description: Full JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401': { description: Invalid code or token }
  /auth/logout:
    post:
      tags: [auth]
      security: [{ adminBearer: [] }]
      summary: Logout (client discards tokens)
      responses:
        '204': { description: No content }
...
  /devices/me/commands/next:
    get:
      tags: [commands]
      security: [{ deviceBearer: [] }]
      summary: Claim next pending command
      responses:
        '200':
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CommandJob'
                  - type: 'null'
        '204': { description: No pending commands }
  /devices/me/commands/{command_id}/complete:
    post:
      tags: [commands]
      security: [{ deviceBearer: [] }]
      parameters:
        - name: command_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandCompleteRequest'
      responses:
        '204': { description: OK }
        '404': { description: Command not found for device }
...
  /devices/{device_id}/commands:
    get:
      tags: [commands]
      security: [{ adminBearer: [] }]
      parameters:
        - name: device_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/CommandJobAdmin' }
                  total: { type: integer }
    post:
      tags: [commands]
      security: [{ adminBearer: [] }]
      parameters:
        - name: device_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommandRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandJobAdmin'
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceBearer:
      type: http
      scheme: bearer
      bearerFormat: DeviceToken
  schemas:
    AdminLoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AdminLoginResponse:
      type: object
      required: [access_token, token_type, requires_2fa]
      properties:
        access_token: { type: string }
        token_type: { type: string, default: bearer }
        requires_2fa: { type: boolean }
    TwoFactorVerifyRequest:
      type: object
      required: [access_token, code]
      properties:
        access_token: { type: string }
        code: { type: string, minLength: 6, maxLength: 8 }
    TokenPair:
      type: object
      required: [access_token, token_type]
      properties:
        access_token: { type: string }
        token_type: { type: string }
...
    CommandType:
      type: string
      enum:
        - ping
        - get_inventory
        - request_upload
        - request_download
        - revoke_local
        - noop
    CommandStatus:
      type: string
      enum: [pending, running, done, failed]
    CreateCommandRequest:
      type: object
      required: [type]
      properties:
        type: { $ref: '#/components/schemas/CommandType' }
        payload:
          type: object
          additionalProperties: true
          nullable: true
    CommandJob:
      type: object
      required: [id, type, payload, status]
      properties:
        id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/CommandType' }
        payload: { type: object, additionalProperties: true, nullable: true }
        status: { $ref: '#/components/schemas/CommandStatus' }
    CommandJobAdmin:
      allOf:
        - $ref: '#/components/schemas/CommandJob'
        - type: object
          properties:
            result: { type: object, additionalProperties: true, nullable: true }
            created_at: { type: string, format: date-time }
            completed_at: { type: string, format: date-time, nullable: true }
    CommandCompleteRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [done, failed]
        result:
          type: object
          additionalProperties: true
          nullable: true
        logs:
          type: string
          maxLength: 65536
          nullable: true
...